Secure variables are null if decryption fails.
At least, it appears there's some authentication as well as decryption with the controlling account's key. It's helpful to know this because one can make steps which depend on a secure variable conditional on that variable being set. If a job with the same appveyor.yml is started under another account that will skip the step rather than having the step fail because e.g. the decryped api key is invalid.